refactor: extract AgentApp and card sections
